Port JKPS to Linux with SFML 3 static build and fix cross-platform bugs
- Fix all SFML 3 API changes: openFromMemory font streaming, scancode
  events, Vector2/FloatRect construction, removed APIs
- Fix segfaults caused by hardcoded oversized sizes passed to
  openFromMemory/loadFromMemory; add _size constants to all embedded
  resources so SFML stream bounds are exact
- Fix #elif linux -> #elif __linux__ for C++17 strict mode (no GNU exts)
- Add missing includes in ParameterLine.cpp (algorithm, Keyboard.hpp)
- Fix CRLF config parsing on Linux so Windows-authored configs load
- Fix hotkey binding for non-US keyboard keys (e.g. German +*~ key)
  by mapping punctuation scancodes directly, bypassing localize/delocalize
- Fix visual key name input to use TextEntered events so any character
  (including +, ö, ü, ä etc.) can be typed regardless of keyboard layout
- Fix non-ASCII visual key display by encoding/decoding via UTF-8
  instead of toAnsiString()
- Bump version to 0.6.0

+++ Source/StatisticLinesPositioner.cpp
@@ -28,9 +28,9 @@ void StatisticsPositioner::operator()()
 
     const auto advMode = Settings::StatisticsTextAdvancedMode;
 
-    const auto width = Settings::StatisticsTextPosition.x + Application::getWindowWidth() - Settings::WindowBonusSizeRight;
+    const auto width = Settings::StatisticsTextPosition.x + static_cast<float>(Application::getWindowWidth()) - static_cast<float>(Settings::WindowBonusSizeRight);
     const auto totalHeight = getTotalHeight();
-    const auto startHeight = Settings::WindowBonusSizeTop + Settings::GfxButtonTextureSize.y / 2.f - totalHeight / 2.f - Settings::StatisticsTextPosition.y;
+    const auto startHeight = static_cast<float>(Settings::WindowBonusSizeTop) + static_cast<float>(Settings::GfxButtonTextureSize.y) / 2.f - totalHeight / 2.f - Settings::StatisticsTextPosition.y;
     auto currentHeight = 0.f;
 
     auto idx = 0u;
